p4i integration Integrate 100039 from 3 to main: Integrate 100038 from 3.1 to 3: Unwarn, kudos to Espen for letting me know about it
[git-p4: depot-paths = "//depot/qt/main/": change = 100041]

+++ src/kernel/qimage.cpp
@@ -4962,7 +4962,7 @@ bool qt_write_dib( QDataStream& s, QImage image )
 		p++;
 	    }
 	}
-	if ( bpl_bmp != (uint)d->writeBlock( (char*)buf, bpl_bmp ) ) {
+	if ( bpl_bmp != d->writeBlock( (char*)buf, bpl_bmp ) ) {
 	    delete[] buf;
 	    return FALSE;
 	}
